refactor(guest-mock-claude): split jsonl integration tests
